api/requirements.txt
anyio==3.6.1
click==8.1.3
fastapi==0.85.0
h11==0.14.0
idna==3.4
joblib==1.2.0
numpy==1.23.3
pandas==1.5.0
pydantic==1.10.2
python-dateutil==2.8.2
pytz==2022.4
scikit-learn==1.1.2
scipy==1.9.2
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
threadpoolctl==3.1.0
typing-extensions==4.4.0
uvicorn==0.18.3
watchfiles==0.17.0
prometheus-client
pyarrow
gitpython